requirements.txt
langchain
langchain-core
langchain-community
pypdf
pymupdf
sentence_transformers
faiss-cpu
chromadb
langchain-groq
python-dotenv